getBook Update
This fixes a good portion of the previous setup in order to properly extract info on a book from goodreads. Some attributes returned for getBook have been removed if the information could no longer be extracted while others have been added as new. A future commit on the documentation will explain everything.